+++ src/main/kotlin/cloud/glitchdev/rfu/gui/window/PartyFinderWindow.kt
@@ -211,7 +211,7 @@ object PartyFinderWindow : BaseWindow(false), Feature {
             height = 100.percent - 5.pixels
         } childOf header
 
-        val createImage = UIImage.ofResource("/assets/rfu/ui/edit.png")
+        val createImage = UIImage.ofResourceCached("/assets/rfu/ui/edit.png")
         UIButton.withImage(createImage, 5f) {
             if (creationOpen && ::creationArea.isInitialized) {
                 creationArea.saveSessionState()
@@ -228,7 +228,7 @@ object PartyFinderWindow : BaseWindow(false), Feature {
             hoverColor = UIScheme.pfInputBgHovered.toConstraint()
         } childOf rightArea
 
-        val filterImage = UIImage.ofResource("/assets/rfu/ui/filter.png")
+        val filterImage = UIImage.ofResourceCached("/assets/rfu/ui/filter.png")
         filterButton = UIButton.withImage(filterImage, 5f) {
             filtersOpen = !filtersOpen
             onUpdate()
@@ -242,7 +242,7 @@ object PartyFinderWindow : BaseWindow(false), Feature {
             hoverColor = UIScheme.pfInputBgHovered.toConstraint()
         } childOf rightArea
 
-        val refreshImage = UIImage.ofResource("/assets/rfu/ui/refresh.png")
+        val refreshImage = UIImage.ofResourceCached("/assets/rfu/ui/refresh.png")
         refreshButton = UIButton.withImage(refreshImage, 5f) {
             PartyWebSocket.syncParties()
             refreshButton.disabled = true